requirements.txt
aniso8601>=9.0.1
anyio>=3.6.1
appdirs>=1.4.4
async-generator>=1.10
attrs>=21.4.0
black>=22.6.0
certifi>=2022.6.15
cffi>=1.15.1
chardet>=5.0.0
charset-normalizer>=2.0.12
click>=8.1.3
colorama>=0.4.5
cryptography>=37.0.4
curlify>=2.2.1
decorator>=5.1.1
exceptiongroup>=1.0.0rc8
Flask>=2.1.3
Flask-RESTful>=0.3.9
flask-restx>=0.5.1
graphql-core>=3.2.1
h11>=0.13.0
hypothesis>=6.53.0
hypothesis-graphql>=0.9.0
hypothesis-jsonschema>=0.22.0
idna>=3.3
iniconfig>=1.1.1
itsdangerous>=2.1.2
Jinja2>=3.1.2
jsonschema>=4.7.2
junit-xml>=1.9
MarkupSafe>=2.1.1
multidict>=6.0.2
mypy-extensions>=0.4.3
numpy>=1.23.1
outcome>=1.2.0
packaging>=21.3
pandas>=1.4.3
pathspec>=0.9.0
platformdirs>=2.5.2
pluggy>=1.0.0
py>=1.11.0
pycparser>=2.21
pyOpenSSL>=22.0.0
pyparsing>=3.0.9
pyrsistent>=0.18.1
PySocks>=1.7.1
pytest>=7.1.2
pytest-subtests>=0.7.0
python-dateutil>=2.8.2
python-dotenv>=0.20.0
pytz>=2022.1
PyYAML>=6.0
regex>=2022.7.24
requests>=2.27.1
retry>=0.9.2
schemathesis>=3.16.0
selenium>=4.3.0
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
starlette>=0.20.4
toml>=0.10.2
tomli>=2.0.1
tomli_w>=1.0.0
trio>=0.21.0
trio-websocket>=0.9.2
typing_extensions>=4.3.0
urllib3>=1.26.10
Werkzeug>=2.1.2
wsproto>=1.1.0
yarl>=1.7.2
